submissions/2026-04-29__001_synthetic_mine_throughput__claude-code__claude-opus-4-7__agent-teams-nelson-max-thinking/requirements.txt
simpy>=4.1.1
numpy>=1.26
pandas>=2.0
scipy>=1.11
networkx>=3.2
pyyaml>=6.0
matplotlib>=3.8
